Calling svauto-deployments.sh via Cloud Init
Now, the demo-stack.yaml calls SVAuto to bootstrap Ubuntu by running Ansible
locally.

This is a good example of how an Instance can pull things from a "config
repository".

Today, this "config repo" is hosted on Github but, in future versions, the
SVAuto host itself, will act as a Git Server for isolated deployments.