SVAuto 0.3.1, introducing openstack.py
We're about to do a complete move to Ansible's Dynamic Inventory for when with
OpenStack deployments.

This first move, only helped us to simplify svauto.sh, later, we'll need a new
version of openstack.py, that supports our Ansible's Hosts Groups on its
Dynamic Inventory engine.

At least, we're getting rid of ugly Bash code, in favor of a much more dynamic
and organized Python code, that talks directly with OpenStack APIs.