Small updates on dpdk-igb-uio-dkms role
Moving the download dir to /root, instead of /tmp.

Also, only build it when the distro is RedHat family, version 7. On CentOS 6,
the code will be there, but not compiled during the Packer build, because
DPDK igb_uio only build on newer kernels, that is only available after rebooting
the box. So, for CentOS 6 PTS, there is a need to manually build the DKMS module
on first boot.

Since PTS on CentOS 6 is not officially released by Sandvine, we can keep this
as a manual process.