New Packer Ubuntu basic templates and seeds!
Now, SVAuto have a solid base to build Ubuntu images (qcows).

Both Ubuntu 16.04 and 14.04 are supported! There will be support for every LTS.

CentOS 6 and 7 are also supported.

To experiment with this, basically, you have to:

1- Bootstrap your Ubuntu Server for SVAuto, run:

  cd ~/svauto
  ./scripts/preset-svauto-server.sh

2- Test the foundations to use Packer with SVAuto:

  cd ~/svauto

  packer build -only=qemu packer/ubuntu16.yaml

  packer build -only=qemu packer/ubuntu14.yaml

  packer build -only=qemu packer/centos7.yaml

  packer build -only=qemu packer/centos6.yaml

That's it!

You'll have RAW images (not qcows for now), under packer/ subdir, for example:

  ls packer/ubuntu16-tmpl/

  ubuntu-16.04.2-server-amd64.raw